Resolve CodeQL code-scanning alerts: path-traversal, ReDoS, XXE, info-exposure hardening (TODO-235)
- Path-injection: route juneau-sc GetConfiguration config lookups through FileUtils.resolveSafely and
  map traversal escape to 403 in LoadConfigResource; add PARSE-method traversal regression test for
  LogsResource.
- ReDoS: fix polynomial-time regex in UrlPathMatcher, bound regex-validation input length in
  HttpPartSchema, and properly escape literal fragments in LogEntryFormatter's date/format templates.
- XXE: disable DTD processing on XmlReader's non-validating path and harden the integration-test
  XmlValidatorParser; add XXE regression test.
- Info exposure: AuthFilter returns a generic Unauthorized body (detail logged server-side);
  PlainTextPojoProcessor emits generic status text for Throwables in production (gated by
  renderResponseStackTraces).
- Correctness: explicit narrowing cast in CharSequenceReader.

Also dismissed 22 by-design/false-positive alerts (SSRF client transports, MsgPack spec narrowing,
utility-level XSS/error-exposure FPs, and already-mitigated log-resource path-injection) with
per-alert rationale recorded on GitHub.

+++ juneau-core/juneau-commons/src/main/java/org/apache/juneau/commons/io/CharSequenceReader.java
@@ -174,7 +174,7 @@ public class CharSequenceReader extends BufferedReader {
 			return 0;
 		long n = Math.min((long) length - next, ns);
 		n = Math.max(-next, n);
-		next += n;
+		next += (int) n;  // Safe narrowing: n is bounded to [-next, length-next], so next+n stays within int range [0, length].
 		return n;
 	}
